-  delete ResultSelfieActivity.kt (move to SelfieActivity.kt) - make responsif
