+++ .idea/deploymentTargetDropDown.xml
@@ -6,12 +6,12 @@
         <type value="RUNNING_DEVICE_TARGET" />
         <deviceKey>
           <Key>
-            <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\ASUS\.android\avd\Pixel_4_API_30_1.avd" />
+            <type value="SERIAL_NUMBER" />
+            <value value="adb-RF8M30PZNDL-po8qu6._adb-tls-connect._tcp" />
           </Key>
         </deviceKey>
       </Target>
     </runningDeviceTargetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-05-23T09:55:43.558966500Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-05-23T14:58:13.729217800Z" />
   </component>
 </project>
