--- app/src/main/java/com/bangkit/ocr_c22_ky03/ResultSelfieActivity.kt
@@ -1,107 +1,0 @@
-package com.bangkit.ocr_c22_ky03
-
-import android.Manifest
-import android.content.Intent
-import android.content.pm.PackageManager
-import android.graphics.Bitmap
-import android.graphics.BitmapFactory
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.widget.Toast
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
-import com.bangkit.ocr_c22_ky03.databinding.ActivityResultSelfieBinding
-import com.bangkit.ocr_c22_ky03.databinding.ActivityTakeSelfieBinding
-import com.bumptech.glide.Glide
-import java.io.File
-import kotlin.math.min
-
-class ResultSelfieActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityResultSelfieBinding
-
-    private var getFile: File? = null
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (requestCode == REQUEST_CODE_PERMISSIONS) {
-            if (!allPermissionsGranted()) {
-                Toast.makeText(
-                    this,
-                    "Tidak mendapatkan permission.",
-                    Toast.LENGTH_SHORT
-                ).show()
-                finish()
-            }
-        }
-    }
-
-    private fun allPermissionsGranted() = REQUIRED_PERMISSIONS.all {
-        ContextCompat.checkSelfPermission(baseContext, it) == PackageManager.PERMISSION_GRANTED
-    }
-
-
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityResultSelfieBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-
-        if (!allPermissionsGranted()) {
-            ActivityCompat.requestPermissions(
-                this,
-                REQUIRED_PERMISSIONS,
-                REQUEST_CODE_PERMISSIONS
-            )
-        }
-
-        binding.captureImage.setOnClickListener { startCameraX() }
-        binding.btnNext.setOnClickListener {
-            startActivity(Intent(this, HistoryActivity::class.java))
-            finish()
-        }
-    }
-
-    private fun startCameraX() {
-        val intent = Intent(this, TakeSelfieActivity::class.java)
-        launcherIntentCameraX.launch(intent)
-    }
-
-    private val launcherIntentCameraX = registerForActivityResult(
-        ActivityResultContracts.StartActivityForResult()
-    ) {
-        if (it.resultCode == CAMERA_X_RESULT) {
-            val myFile = it.data?.getSerializableExtra("picture") as File
-            val isBackCamera = it.data?.getBooleanExtra("isBackCamera", true) as Boolean
-
-            val result =
-                rotateBitmap(
-                    BitmapFactory.decodeFile(myFile.path),
-                    isBackCamera
-                )
-            val resultCircle = result.toSelfie(result)
-            getFile = bitmapToFile(resultCircle, application)
-            Glide.with(this)
-                .load(resultCircle)
-                .into(binding.previewImageView)
-
-
-//            binding.previewImageView.setImageBitmap(result)
-        }
-
-    }
-
-
-    companion object {
-        const val CAMERA_X_RESULT = 200
-
-        private val REQUIRED_PERMISSIONS = arrayOf(Manifest.permission.CAMERA)
-        private const val REQUEST_CODE_PERMISSIONS = 10
-    }
-}
