+++ app/src/main/java/com/bangkit/ocr_c22_ky03/Utils.kt
@@ -119,7 +119,7 @@ fun bitmapToFile(bitmap: Bitmap?, application: Application): File {
     val photoFile = createFile(application)
     val bmpStream = ByteArrayOutputStream()
 
-    bitmap?.compress(Bitmap.CompressFormat.PNG, 100, bmpStream)
+    bitmap?.compress(Bitmap.CompressFormat.JPEG, 100, bmpStream)
     val bmpPicByteArray = bmpStream.toByteArray()
 
     val fos = FileOutputStream(photoFile)
