+++ app/src/main/res/layout/activity_take_selfie.xml
@@ -26,16 +26,31 @@
 
     <ImageView
         android:id="@+id/img_border"
-        android:layout_width="349dp"
-        android:layout_height="545dp"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
         android:layout_marginStart="16dp"
+        android:layout_marginTop="32dp"
         android:layout_marginEnd="16dp"
-        app:layout_constraintBottom_toTopOf="@+id/captureImage"
+        android:layout_marginBottom="16dp"
+        app:layout_constraintBottom_toTopOf="@+id/tv_messege"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="@+id/viewFinder"
-        app:layout_constraintVertical_bias="0.71000004"
         app:srcCompat="@drawable/oval_1" />
+
+    <TextView
+        android:id="@+id/tv_messege"
+        style="@style/Text.Description"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="16dp"
+        android:layout_marginEnd="16dp"
+        android:layout_marginBottom="8dp"
+        android:text="Position your face in the Frame!"
+        android:textColor="@color/white"
+        app:layout_constraintBottom_toTopOf="@+id/captureImage"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
